+++ src/main/java/cc/cassian/campfire/mixin/CampfireBlockEntityMixin.java
@@ -1,0 +1,37 @@
+package cc.cassian.campfire.mixin;
+
+import net.minecraft.block.BlockState;
+import net.minecraft.block.entity.CampfireBlockEntity;
+import net.minecraft.entity.effect.StatusEffect;
+
+//? if >1.21.2 {
+import net.minecraft.recipe.ServerRecipeManager;
+import net.minecraft.recipe.input.SingleStackRecipeInput;
+import net.minecraft.registry.entry.RegistryEntry;
+import net.minecraft.server.world.ServerWorld;
+import net.minecraft.recipe.CampfireCookingRecipe;
+//?}
+
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import static cc.cassian.campfire.CampfireMod.applyPlayerEffects;
+
+@Mixin(CampfireBlockEntity.class)
+public class CampfireBlockEntityMixin {
+    @Inject(method = "litServerTick", at = @At(value = "HEAD"))
+    //? if >1.21.2 {
+    private static void mixin(ServerWorld world, BlockPos pos, BlockState state, CampfireBlockEntity blockEntity, ServerRecipeManager.MatchGetter<SingleStackRecipeInput, CampfireCookingRecipe> recipeMatchGetter, CallbackInfo ci)
+
+        //?} else {
+        /*private static void mixin(World world, BlockPos pos, BlockState state, CampfireBlockEntity campfire, CallbackInfo ci)
+         *///?}
+    {
+        applyPlayerEffects(world, pos);
+    }
+
+}