+++ CHANGELOG.md
@@ -4,15 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Config to set amplifier of effect.
+
+### Changed
+- Migrated buildscript to Stonecutter, supporting 1.19.2, 1.20.1, and 1.21.1.
+
 ## [1.1.0] - 2024-01-06
 
-## Added
+### Added
 - Config options to adjust distance, duration, and whether to use Regeneration or Comfort.
 
-## Changed
+### Changed
 - Lowered default duration from 90 seconds to 5 seconds.
 
-## Fixed
+### Fixed
 - Campfire Y distance is now properly capped.
 
 ## [1.0.0] - 2024-01-05